Feature: add display information by id
* fix: resolve test conflict

* fix: change detail text

* feat: show labor registration by id

* fix: update image url

* docs: update doc

* merge: resolve conflict

+++ app/Http/Controllers/LaborRegistrationController.php
@@ -204,7 +204,7 @@ class LaborRegistrationController extends Controller
         ], 201);
     }
 
-     /**
+    /**
      * Retrieve a paginated list of labor registrations (elves).
      *
      * @param  \Illuminate\Http\Request $request The HTTP request object.
@@ -325,6 +325,131 @@ class LaborRegistrationController extends Controller
         return response()->json($data, 200);
     }
 
+
+    /**
+     * Retrieve a labor registrations (elves).
+     *
+     * @param  int $id The id of the labor registration.
+     * @return \Illuminate\Http\JsonResponse The paginated list of labor registrations.
+     * 
+     * @OA\Get(
+     *     path="/v1/labor-registration/{id}",
+     *     summary="Retrieve a labor registrations (elves)",
+     *     tags={"Labor-Registration"},
+     *     @OA\Parameter(
+     *          name="X-API-Key",
+     *          in="header",
+     *          description="API Key",
+     *          @OA\Schema(
+     *              type="string"
+     *          )
+     *      ),
+     *     @OA\Parameter(
+     *         name="id",
+     *         in="path",
+     *         description="ID of the labor registration",
+     *         required=true,
+     *         @OA\Schema(
+     *             type="integer",
+     *             example=1
+     *         )
+     *     ),
+     *     @OA\Response(
+     *         response=200,
+     *         description="Successful operation",
+     *         @OA\JsonContent(
+     *             type="object",
+     *             @OA\Property(
+     *                 property="data",
+     *                 type="array",
+     *                 @OA\Items(
+     *                     @OA\Property(property="image", type="string", example="http://localhost:8000/storage/image/image.jpg"),
+     *                     @OA\Property(property="name", type="string", example="John Doe"),
+     *                     @OA\Property(property="email", type="string", example="example@example.com"),
+     *                     @OA\Property(property="age", type="integer", example=25),
+     *                     @OA\Property(property="address", type="string", example="1234 Main St"),
+     *                     @OA\Property(property="height", type="number", example=1.75),
+     *                     @OA\Property(property="created_at", type="string", example="2023-05-01 12:00:00"),
+     *                     @OA\Property(property="updated_at", type="string", example="2023-05-01 12:00:00"),
+     *                     @OA\Property(property="id", type="integer", example=1)
+     *                 )
+     *             ),
+     *         )
+     *     ),
+     *     @OA\Response(
+     *         response=404,
+     *         description="Resource not found",
+     *         @OA\JsonContent(
+     *             @OA\Property(
+     *                 property="errors",
+     *                 type="object",
+     *                 @OA\Property(
+     *                     property="status",
+     *                     type="integer",
+     *                     example=404
+     *                 ),
+     *                 @OA\Property(
+     *                     property="title",
+     *                     type="string",
+     *                     example="Not Found"
+     *                 ),
+     *                 @OA\Property(
+     *                     property="detail",
+     *                     type="string",
+     *                     example="Resource not found."
+     *                 )
+     *             )
+     *          )
+     *     ),
+     *     @OA\Response(
+     *         response=406,
+     *         description="Invalid ID",
+     *         @OA\JsonContent(
+     *             type="object",
+     *             @OA\Property(property="errors", type="array",
+     *                 @OA\Items(
+     *                     type="object",
+     *                     @OA\Property(property="status", type="integer", example=406),
+     *                     @OA\Property(property="title", type="string", example="Invalid Id"),
+     *                     @OA\Property(property="detail", type="string", example="Provided Id didn’t match standard.")
+     *                 )
+     *             )
+     *         )
+     *     )
+     * )
+     */
+    public function show($id): JsonResponse
+    {
+        if (!is_numeric($id) && $id < 1) {
+            throw new InvalidIdException();
+        }
+
+        $data = LaborRegistration::find($id);
+
+        if (!$data) {
+            throw new LaborRegistrationNotFoundException();
+        }
+
+        $urlImage = url('storage/image/' . $data->image);
+
+        return response()->json([
+            'data' => [
+                [
+                    'image' => $urlImage,
+                    'name' => $data->name,
+                    'email' => $data->email,
+                    'age' => $data->age,
+                    'address' => $data->address,
+                    'height' => $data->height,
+                    'created_at' => $data->created_at,
+                    'updated_at' => $data->updated_at,
+                    'id' => $data->id,
+                ]
+            ]
+        ], 200);
+    }
+
+
     /**
      * @OA\Delete(
      *     path="/labor-registration/{id}",
@@ -419,25 +544,29 @@ class LaborRegistrationController extends Controller
      *         )
      *     )
      * )
-    */
+     */
     public function destroy(string $id)
     {
-        if( !is_numeric($id) ) {
+        if (!is_numeric($id)) {
             throw new InvalidIdException();
         }
 
         $data = LaborRegistration::find($id);
 
+        if (!$data) {
+            throw new LaborRegistrationNotFoundException();
+        }
+
+
         if (is_null($data)) {
             throw new LaborNotFoundException();
         }
-    
+
         $data->delete();
-    
+
         return response()->json([
             'message' => 'Elf successfully deleted.',
             'data' => $data
         ], 200);
     }
 }
-
